init Employee

+++ database/migrations/2022_05_29_144519_create_employees_table.php
@@ -17,7 +17,7 @@ class CreateEmployeesTable extends Migration
             $table->id();
             $table->string('first_name');
             $table->string('last_name');
-            $table->integer('cpf')->unique();
+            $table->bigInteger('cpf')->unique();
             $table->foreignId('user_id')
                 ->constrained('users')
                 ->onUpdate('cascade')
